% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{orderly_remote_sharedrive}
\alias{orderly_remote_sharedrive}
\title{Create an orderly remote based at a path}
\usage{
orderly_remote_sharedrive(path, name = NULL)
}
\arguments{
\item{path}{Path to use as a remote}

\item{name}{Friendly name for the remote}
}
\value{
An \code{orderly_remote_sharedrive} object

An \code{orderly_remote_sharedrive} object, designed to be
  used by orderly.  This function should however not generally be
  called by users directly, as it should be used within
  \code{orderly_config.yml}
}
\description{
Implements an orderly "remote" using a shared drive as a backend.  Use
this within an \code{orderly_config.yml} configuration.
}
\details{
A configuration might look like:

\preformatted{
remote:
  real:
    driver: orderly.sharedrive::orderly_remote_sharedrive
    args:
      path: ~/path/to/network/drive
}

which would create a remote called \code{real}, at the file path
specified. This can be a network drive or a one drive synced drive.

This function is not intended to be used interactively
}
