% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{orderly_remote_sharepoint}
\alias{orderly_remote_sharepoint}
\title{Create an orderly remote based on Sharepoint}
\usage{
orderly_remote_sharepoint(url, site, path, name = NULL)
}
\arguments{
\item{url}{Sharepoint URL}

\item{site}{Sharepoint "site"}

\item{path}{Path within the Sharepoint site. In our experience
these often start with \code{Shared Documents} but your setup
may vary.}

\item{name}{Friendly name for the remote}
}
\value{
An \code{orderly_remote_sharepoint} object

An \code{orderly_remote_sharepoint} object, designed to be
  used by orderly.  This function should however not generally be
  called by users directly, as it should be used within
  \code{orderly_config.yml}
}
\description{
Implements an orderly "remote" using Sharepoint as a backend.  Use
this within an \code{orderly_config.yml} configuration.
}
\details{
A configuration might look like:

\preformatted{
remote:
  real:
    driver: orderly.sharepoint::orderly_remote_sharepoint
    args:
      url: https://example.sharepoint.com
      site: mysite
      path: Shared Documents/orderly/real
}

which would create a remote called \code{real}, using your group's
Sharepoint hosted at \code{https://example.sharepoint.com}, on
site \code{mysite} and within that site using path \code{Shared
Documents/orderly/real}.

Currently authentication is interactive, or uses the values of
environment variables \code{SHAREPOINT_USERNAME} and
\code{SHAREPOINT_PASS}.  Once we expose richer authentication
approaches in spud that will be exposed here (RESIDE-162).

This function is not intended to be used interactively
}
